+++ prod/Watchdogd/Include/Zmq.h
@@ -48,7 +48,7 @@
  extern gboolean Zmq_Send_as_raw ( struct ZMQUEUE *zmq, void *buf, gint taille );
  extern gint Recv_zmq ( struct ZMQUEUE *zmq, void *buf, gint taille_buf );
  extern void Zmq_Send_DI_to_master ( struct LIBRAIRIE *lib, gchar *tech_id, gchar *acronyme, gboolean etat );
- extern void Zmq_Send_AI_to_master ( struct LIBRAIRIE *lib, gchar *tech_id, gchar *acronyme, gfloat valeur, gboolean in_range);
+ extern void Zmq_Send_AI_to_master ( struct LIBRAIRIE *lib, gchar *tech_id, gchar *acronyme, gdouble valeur, gboolean in_range);
  extern void Zmq_Send_CDE_to_master ( struct LIBRAIRIE *lib, gchar *tech_id, gchar *acronyme );
  extern void Zmq_Send_WATCHDOG_to_master ( struct LIBRAIRIE *lib, gchar *tech_id, gchar *acronyme, gint consigne );
  extern JsonNode *Recv_zmq_with_json ( struct ZMQUEUE *zmq, const gchar *my_tech_id, gchar *buf, gint taille_buf );
